refactor : add list api using dto & jpa
dto와 jpa를 사용한 상품 조회 인수 테스트 및 api 생성
