+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -2,10 +2,10 @@ package kitchenpos.menu.application;
 
 import kitchenpos.menu.dao.MenuDao;
 import kitchenpos.menugroup.dao.MenuGroupDao;
-import kitchenpos.dao.MenuProductDao;
+import kitchenpos.menu.dao.MenuProductDao;
 import kitchenpos.product.dao.ProductDao;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.product.domain.Product;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
